split into smaller files; add tests
